src/main_service/server/requirements.txt
Flask>=3.0.2
pyjwt==2.8.0
cryptography==41.0.7
requests>=2.31.0
grpcio==1.40.0
protobuf==3.20
kafka-python==2.0.2